requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
async-timeout==5.0.1
attrs==25.4.0
colorama==0.4.6
discord.py==2.6.4
disnake==2.7.0
disnake-pagination==2.1
frozenlist==1.8.0
idna==3.11
mpmath==1.4.0
multidict==6.7.1
pillow==12.1.1
propcache==0.4.1
python-dotenv==1.2.1
six==1.17.0
topggpy==1.4.0
trueskill==0.4.5
typing_extensions==4.15.0
websockets==16.0
yarl==1.22.0
